+++ src/components/common/PlanInput.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { useEffect, useState } from "react";
 import { PlaneIcon } from "../../assets/svgs";
 
@@ -12,7 +14,7 @@ const PlanInput = () => {
   }, [title])
 
   return (
-    <div className={`flex justify-between gap-2.5 rounded-full px-6 py-4.5 bg-white border border-gray-300`}>
+    <div className={`flex justify-between gap-2.5 rounded-full w-full px-6 py-4.5 bg-white border border-gray-300 overflow-hidden`}>
       <input 
         className={`text-[20px] font-light text-gray-700 placeholder:text-gray-300 grow`}
         placeholder="여행 플랜 이름을 입력해주세요"
