+++ src/components/plan/map/PlaceList.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import Toggle from "@/component/common/Toggle";
+import Toggle from "@/components/common/Toggle";
 import { PlaceProps } from "@/types/Plan/PlaceType";
 import { useState } from "react";
 import Place from "./Place";
