+++ src/constants/Tag.ts
@@ -1,0 +1,27 @@
+export const TAGLIST = [
+  {
+    title: "지역",
+    tagList: [
+      "속초",
+      "부산",
+      "경주",
+      "제주도",
+    ]
+  },
+  {
+    title: "누구",
+    tagList: [
+      "가족",
+      "친구",
+      "연인",
+    ]
+  },
+  {
+    title: "테마",
+    tagList: [
+      "맛집 탐방",
+      "힐링 여행",
+      "가족 여행",
+    ]
+  },
+]
