+++ src/stories/plan/map/PlaceList.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import PlaceList from '@/component/plan/map/PlaceList';
+import PlaceList from '@/components/plan/map/PlaceList';
 import { PlaceItems } from '@/mocks/places';
 
 const meta = {
