+++ src/stories/plan/schedule/ScheduleBox.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import ScheduleBox from '@/component/plan/schedule/ScheduleBox';
+import ScheduleBox from '@/components/plan/schedule/ScheduleBox';
 
 const meta = {
   title: 'Plan/Schedule/ScheduleBox',
